Moving filter to isolated file
